requirements.txt
fireworks-ai==0.19.18
numpy==2.3.2
openai_harmony==0.0.4
PyYAML==6.0.2
scipy==1.16.1
pandas>=1.3.0
importlib_resources>=1.3.0; python_version < "3.9"
